container/requirements.txt
numpy==2.3.1 # numpy 最新版是2.3.1，最大支持python3.12
onnxruntime==1.22.0
pillow==11.2.1
# uvicorn==0.34.3
# starlette==0.47.1
aiohttp==3.11.18
